Add blueprints
