+++ jsonschema_scratch/config.py
@@ -1,0 +1,2 @@
+SCHEMAS_DIR_NAME = "schemas"
+SCHEMAS_RELATIVE_PATH = "../schemas"
